@@ -3,10 +3,10 @@ description: Demonstrates how to use the image_watermark plugin.
 
 publish_to: 'none' 
 
-version: 1.0.0+2
+version: 1.0.1+3
 
 environment:
-  sdk: '>=2.17.3 <3.0.0'
+  sdk: '>=3.3.1 <4.0.0'
 
 dependencies:
   flutter:
@@ -18,7 +18,7 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
     path: ../
-  image_picker: ^0.8.6
+  image_picker: ^1.0.7
 
   cupertino_icons: ^1.0.2
 
@@ -26,7 +26,10 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
 
 flutter:
   uses-material-design: true
+
+  assets:
+  - assets/fonts/file.zip
